Container v1: Move testing files to testing directory
